split fixes
module-split rule
